% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{calc_dry_days}
\alias{calc_dry_days}
\title{Calculate dry days from everwader}
\usage{
calc_dry_days(depth_data)
}
\arguments{
\item{depth_data}{depth .nc files}
}
\description{
Calculate dry days from everwader
}
