% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{calc_recession}
\alias{calc_recession}
\title{Calculate recession from everwader}
\usage{
calc_recession(depth_data)
}
\arguments{
\item{depth_data}{depth .nc files}
}
\description{
Calculate recession from everwader
}
