% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{calc_reversals}
\alias{calc_reversals}
\title{Calculate reversals following Peterson 2017}
\usage{
calc_reversals(depth_data)
}
\arguments{
\item{depth_data}{depth .nc files}
}
\description{
Calculate reversals following Peterson 2017
}
