% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorCombos}
\alias{colorCombos}
\title{create all color combinations}
\usage{
colorCombos(v = c("R", "Y", "W", "DB", "G", "O"))
}
\arguments{
\item{v}{character vector containing the colors}
}
\description{
colorCombos
}
\examples{
colorCombos()
}
