% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eden.R
\name{combine_eden_depths}
\alias{combine_eden_depths}
\title{Combine the quarterly EDEN depths .nc files into a single .nc file}
\usage{
combine_eden_depths(
  eden_path = file.path(wader::get_default_data_path(), "EvergladesWadingBird/Water")
)
}
\arguments{
\item{eden_path}{path where the EDEN data should be stored}
}
\value{
char path of combined file
}
\description{
Combine the quarterly EDEN depths .nc files into a single .nc file
}
