% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\name{consecutive_pts_dist}
\alias{consecutive_pts_dist}
\title{consecutive_pts_dist}
\usage{
consecutive_pts_dist(x)
}
\description{
consecutive_pts_dist
}
\examples{
x = fetch_GPS_tracks(all = TRUE, withinStudyArea = TRUE) \%>\%
	   consecutive_pts_dist


}
