% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\name{data.tableTransform}
\alias{data.tableTransform}
\title{projp}
\usage{
data.tableTransform(DT,
  proj = "+proj=omerc +lat_0=71.3332703512554 +lonc=-156.654269449915 +alpha=-41.5 +gamma=0.0 +k=1.000000 +x_0=0.000 +y_0=0.000 +ellps=WGS84 +units=m")
}
\description{
projp
}
