% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_backup}
\alias{db_backup}
\title{database backup}
\usage{
db_backup(year = data.table::year(Sys.Date()), db = yy2dbnam(year),
  host = ip(), user = getOption("wader.user"), pwd = wadeR::pwd(),
  outdir = getOption("wader.dbbackup"), startMonth = 5,
  startDay = 25)
}
\description{
database backup
}
\note{
similar to sdb::mysqldump but tailored for a field database.
}
