% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{discharge}
\alias{discharge}
\title{Calculate stream discharge}
\usage{
discharge(widths, depths, vel)
}
\arguments{
\item{widths}{Transect measurement points}

\item{depths}{Depth measurements}

\item{velocity}{Velocity measurements}
}
\value{
a numeric
}
\description{
Calculate stream discharge
}
