% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eden.R
\name{download_eden_depths}
\alias{download_eden_depths}
\title{Download the EDEN depths data}
\usage{
download_eden_depths(eden_path, force_update = FALSE)
}
\arguments{
\item{eden_path}{path where the EDEN data should be stored}

\item{force_update}{if TRUE update all data files even if checks indicate
that remote files are unchanged since the current local copies were
created}
}
\value{
char vector of downloaded/updated files
}
\description{
Download the EDEN depths data
}
