% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{dry_down}
\alias{dry_down}
\title{Calculate dry down rates from water gauge data}
\usage{
dry_down(
  depths,
  minyear = as.integer(format(Sys.Date(), "\%Y")) - 6,
  maxyear = as.integer(format(Sys.Date(), "\%Y")),
  download_if_missing = TRUE
)
}
\arguments{
\item{depths}{Depth data .csv file}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{download_if_missing}{if the specified file path doesn't have the
EvergladesWadingBird folder, then download it}
}
\value{
a data.frame
}
\description{
Calculate dry down rates from water gauge data, by subregion.
Calculate days from high to high November to January (early) and January to March (late),
calculate max difference relative to number of days. Also calculate exceedance, \% of
previous years with slower dry down rates.
}
