% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{EGGS_CHICKS_updateHatchDate}
\alias{EGGS_CHICKS_updateHatchDate}
\title{EGGS_CHICKS_updateHatchDate}
\usage{
EGGS_CHICKS_updateHatchDate(table = "EGGS_CHICKS", db)
}
\description{
EGGS_CHICKS_updateHatchDate
}
\examples{
EGGS_CHICKS_updateHatchDate()
}
