% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_finder.R
\name{epi_finder}
\alias{epi_finder}
\title{Bottom of epilimnion finder}
\usage{
epi_finder(depths, temps)
}
\arguments{
\item{depths}{Numeric vector of depths corresponding to the temps values}

\item{temps}{Numeric vector of water temperature in degrees Celsius}
}
\description{
Determines the depth of the bottom of the epilimnion in a lake profile.
}
\keyword{thermocline}
