% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{extract_region_means}
\alias{extract_region_means}
\title{Calculate region means from raster data}
\usage{
extract_region_means(raster, regions)
}
\arguments{
\item{raster}{variable raster}

\item{regions}{regions polygons}
}
\value{
region means spdf
}
\description{
Calculate region means from raster data
}
