% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_work.R
\name{foraging_indicator}
\alias{foraging_indicator}
\title{Generate summaries of tactile/visual foraging indicator data}
\usage{
foraging_indicator(
  path = get_default_data_path(),
  minyear = 1986,
  maxyear = 2021,
  window = 3,
  download_if_missing = TRUE
)
}
\arguments{
\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{window}{number of years over which to create a rolling average}
}
\value{
a data.frame
}
\description{
Create a table of rolling averages for the proportion of tactile/visual foragers,
by year
}
