% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eden.R
\name{get_data_urls}
\alias{get_data_urls}
\title{Get EDEN depths data URLs for download}
\usage{
get_data_urls(file_names)
}
\arguments{
\item{file_names}{file names to download from metadata}
}
\value{
list of file urls
}
\description{
Get EDEN depths data URLs for download
}
