% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_eden.R
\name{get_files_to_update}
\alias{get_files_to_update}
\title{Determine list of new EDEN files to download}
\usage{
get_files_to_update(
  eden_path = file.path(get_default_data_path(), "EvergladesWadingBird/Water"),
  metadata,
  force_update = FALSE
)
}
\arguments{
\item{eden_path}{path where the EDEN data should be stored}

\item{metadata}{EDEN file metadata}

\item{force_update}{if TRUE update all data files even if checks indicate
that remote files are unchanged since the current local copies were
created}
}
\description{
Determine list of new EDEN files to download
}
