% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_spatial_units.R
\name{get_state_spatial_units}
\alias{get_state_spatial_units}
\title{get_state_spatial_units}
\usage{
get_state_spatial_units(state, unit_type, year)
}
\arguments{
\item{state}{character name of state}

\item{unit_type}{chr, type of unit to be mapped; acceptable options are "county", "huc", or "aquifer"}

\item{year}{int, a year for which representative units are needed. (not implemented)}
}
\description{
Returns an `SpatialPolygonsDataframe` of county, huc, or aquifer boundaries for a state. (Aquifers not supported yet)
}
\examples{
state <- "Delaware" 
unit_type <- "huc"
year <- 2010
plot_units <- get_state_spatial_units(state, unit_type, year)
plot_units

}
