% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wade_method.R
\name{get_wade_method}
\alias{get_wade_method}
\title{get_wade_method}
\usage{
get_wade_method(wade_method_url)
}
\arguments{
\item{wade_method_url}{character, a url that will return information about methodology used for a specific organization/report}
}
\value{
\code{data.frame} containing methodology information
}
\description{
Get method used for WaDE data.
}
\examples{
wade_method_url <- paste0("https://water.utah.gov/DWRE/WADE/v0.2/GetMethod/GetMethod.php?methodid=utwre&methodname=CONSUMPTIVE_USE")
get_wade_method(wade_method_url)

}
