% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_counts.R
\name{ground_counts}
\alias{ground_counts}
\title{Generate ground count summaries}
\usage{
ground_counts(
  path = get_default_data_path(),
  minyear = 1996,
  maxyear = as.numeric(format(Sys.Date(), "\%Y")),
  level = "region"
)
}
\arguments{
\item{path}{either the file path that contains the Wading Bird folder or
"repo", which then pulls data from the EvergladesWadingBird GitHub repository}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{level}{level at which to summarize (point, region)}
}
\value{
a data.frame
}
\description{
Generate summaries of ground transect count at waypoint or region level
}
