% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{hatch_est}
\alias{hatch_est}
\title{Calculates the estimated hatching date}
\usage{
hatch_est(x, y)
}
\arguments{
\item{x}{a data.table containing
nest,
arrival_datetime (POSIXct),
float_angle (angle with which the egg is floating in degree (must be >20 and <90) ),
float_height (only if egg is floating) float height in mm
pk (row id)}

\item{y}{a data.table containing
nest,
iniCLutch (initial clutch),
clutch (final clutch),
datetime_found}
}
\value{
an estimated hatching date (as POSIXct)
}
\description{
This function calculates the estimated hatching date based on Liebezeit et al. (2007)
              "Assessing the development of shorebird eggs using the flotation method:
              Species-specific and generalized regression models." Condor 109(1): 32-47.
}
\examples{
 #----------------------------------------------------#
x = data.table(
  nest = c('R101', 'N102', 'P101', 'X102', 'R103', 'R104'),
  arrival_datetime =as.POSIXct(c('2017-06-01 10:00','2017-06-03 11:00','2017-06-13 09:00','2017-06-13 09:00',NA,NA)),
  float_angle = c(21, 80, 85, 20,NA,NA),
  float_height = c(0, NA, 2, NA,NA,NA), pk = 1:6 )

y = data.table(nest = c('R103', 'R104'), datetime_ = as.POSIXct(c('2017-06-01 10:00', '2017-06-01 13:00')),
    iniClutch = c(1,2), clutch = c(4,4) )

hatch_est(x,y)
}
