% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{hatch_est_db}
\alias{hatch_est_db}
\title{Estimated hatching date based on Liebezeit (2007) database version}
\usage{
hatch_est_db(species = "REPH", float_time, float_angle, float_height,
  found_datetime = NA, found_clutch = 4, final_clutch = 4)
}
\arguments{
\item{species}{Species (Red Phalarope = 'REPH', Red-necked Phalarope = 'RNPH', Pectoral Sandpiper = 'PESA')}

\item{float_time}{Datetime eggs were floated}

\item{float_angle}{Float angle (must be >= 21 and <89)}

\item{float_height}{If egg breaks the surface give the mm above it}

\item{found_datetime}{Datetime when the nest was found}

\item{found_clutch}{Clutch size when found}

\item{final_clutch}{Clutch size when complete}
}
\value{
Estimated hatching datetime
}
\description{
Estimated hatching date based on Liebezeit (2007) database version
}
\examples{
hatch_est_db(found_datetime = as.POSIXct('2017-06-01 12:00'), found_clutch = 2, final_clutch = 4)
hatch_est_db(float_time = as.POSIXct('2017-06-02 00:00'), float_angle = 30, float_height = NA)
}
