% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{initiation}
\alias{initiation}
\title{Initiation date}
\usage{
initiation(
  found_datetime,
  found_clutch,
  final_clutch,
  hatching_datetime,
  est_hatching_datetime,
  incubation_period,
  incubator = FALSE
)
}
\arguments{
\item{found_datetime}{Datetime when the nest was found}

\item{found_clutch}{Clutch size when found}

\item{final_clutch}{Clutch size when complete}

\item{hatching_datetime}{Datetime hatching (clutch mean)}

\item{est_hatching_datetime}{Estimated hatching time}

\item{incubator}{If eggs where incubated in incubator = TRUE (default = FALSE)}
}
\value{
Nest initiation datetime
}
\description{
Calculate the inititaion date based on
1. initial incomplete clutch size
2. hatching datetime (adjusted if eggs where in incubator)
3. estimated hatching datetime
}
\examples{
initiation(found_datetime = as.POSIXct('2017-06-26 12:00'), found_clutch = 2, final_clutch = 3)
initiation(found_datetime = as.POSIXct('2017-06-06 10:00'), found_clutch = 3, final_clutch = 3,
           hatching_datetime = as.POSIXct('2017-06-26 10:00'))
initiation(found_datetime = as.POSIXct('2017-06-06 10:00'), found_clutch = 3, final_clutch = 3,
           hatching_datetime = as.POSIXct('2017-06-26 10:00'), incubator = TRUE)
initiation(found_datetime = as.POSIXct('2017-06-06 12:00'), found_clutch = 3, final_clutch = 3,
           hatching_datetime = NA, est_hatching_datetime = as.POSIXct('2017-06-26 12:00'))
}
