% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_work.R
\name{initiation_indicator}
\alias{initiation_indicator}
\title{Generate summaries of wood stork nesting intiation indicator data}
\usage{
initiation_indicator(
  path = get_default_data_path(),
  minyear = 1986,
  maxyear = 2021,
  window = 3,
  download_if_missing = TRUE
)
}
\arguments{
\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{window}{number of years over which to create a rolling average}
}
\value{
a data.frame
}
\description{
Create a table of rolling averages for the date of wood stork nest initiation,
by year
}
