% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{inspector}
\alias{inspector}
\title{Inspector functions}
\description{
inspector functions as defined in DataEntry.validation package
}
\examples{
\dontrun{
x = idbq('select * from EGGS_CHICKS ')
class(x) = c(class(x), 'EGGS_CHICKS')
ii = inspector(x)
evalidators(ii)
}

}
