% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ip}
\alias{ip}
\title{Find the ip address.}
\usage{
ip()
}
\value{
The host external ip. It falls back to localhost on error.
}
\description{
Find the ip address.
}
\note{
The function is used by idbq and in global.R files
}
