% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{load_boundaries}
\alias{load_boundaries}
\title{Calculate dry days from everwader}
\usage{
load_boundaries(
  path = file.path(wader::get_default_data_path(),
    "EvergladesWadingBird/SiteandMethods/regions"),
  level = "subregions"
)
}
\arguments{
\item{path}{path to regions shapefile}

\item{level}{region level to load (all, wcas, or subregions)}
}
\description{
Calculate dry days from everwader
}
