% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_elements.R
\name{maps}
\alias{maps}
\alias{theme_wader}
\alias{coord_wader}
\alias{map_empty}
\alias{map_nests}
\alias{map_tracks}
\alias{map_resights}
\alias{map_resights_by_id}
\alias{map_resights_by_day}
\title{maps}
\usage{
theme_wader()

coord_wader(x = map_layers)

map_empty(m = map_layers)

map_nests(n, size = 2.5)

map_tracks(x = fetch_GPS_tracks(48))

map_resights(x, size = 2.5)

map_resights_by_id(x, size = 2.5)

map_resights_by_day(x, n, size = 2.5)
}
\examples{
x = map_layers
coord_wader(x)
map_empty()
n = NESTS()[species == 'REPH']
map_nests(n)
map_tracks()
x = RESIGHTINGS()
map_resights(x)
map_resights(x, zoom = FALSE)
x = RESIGHTINGS_BY_ID('DB,Y', 'Y')
map_resights_by_id(x)

x = RESIGHTINGS_BY_DAY()
n = NESTS()
n = n[nest2species(nest) == 'REPH']
map_resights_by_day(x)
map_resights_by_day(x, n)

}
