% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_work.R
\name{max_count_plot}
\alias{max_count_plot}
\title{Generate summaries of max count data for target species and plot}
\usage{
max_count_plot(
  path = get_default_data_path(),
  minyear = 1980,
  maxyear = 2021,
  window = 3,
  download_if_missing = TRUE
)
}
\arguments{
\item{path}{either the file path that contains the Wading Bird folder or
"repo", which then pulls data from the EvergladesWadingBird GitHub repository}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{window}{number of years over which to create a rolling average}

\item{download_if_missing}{if the specified file path doesn't have the
EvergladesWadingBird folder, then download it}
}
\value{
a data.frame
}
\description{
Create a table of rolling averages for the max count indicator data,
by year and region
}
