% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_counts.R
\name{max_counts}
\alias{max_counts}
\title{Generate max count summaries}
\usage{
max_counts(
  path = get_default_data_path(),
  minyear = 1986,
  maxyear = format(Sys.Date(), "\%Y"),
  level = "colony"
)
}
\arguments{
\item{path}{either the file path that contains the Wading Bird folder or
"repo", which then pulls data from the EvergladesWadingBird GitHub repository}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{level}{level at which to summarize (colony, subregion, region, all)}
}
\value{
a data.frame
}
\description{
Generate summaries of max count at colony, subregion or region level
}
