% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalimnion_finder.R
\name{metalimnion_finder}
\alias{metalimnion_finder}
\title{Metalimnion finder}
\usage{
metalimnion_finder(temps, depths, slope = 0.1, seasonal = TRUE,
  mixed_cutoff = 1)
}
\arguments{
\item{temps}{Vector of water temperature in degrees Celsius}

\item{depths}{Vector of depths corresponding to the temps values}

\item{seasonal}{Logical value indicating if seasonal thermocline should be returned.
The seasonal thermocline is defined as the deepest density gradient found in the profile.}

\item{mixed_cutoff}{Cutoff point where the thermocline is not calculated; defaults to 1 degree C}
}
\description{
Determines the upper and lower bounds of the metalimnion in a water column from a temperature profile of a stratified lake.
This function was borrowed heavily from the rLakeAnalyzer package on CRAN.
}
\examples{

depths <- c(0, 1, 2, 3, 4, 5, 6, 7, 8)
temps <- c(25, 25, 25, 20, 15, 12, 10, 8, 8, 8)

metalimnion_finder(temps, depths)
}
\keyword{thermocline}
