% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fielddb_updates.R
\name{nanotag2db}
\alias{nanotag2db}
\title{nanotag txt to database}
\usage{
nanotag2db(dir = "~/ownCloud/RAW_DATA/NANO_TAG_DATA/")
}
\value{
TRUE if new data are updated, error on no new files
}
\examples{
\dontrun{
 require(wader)

 nanotag2db()
}

}
\author{
MV
}
