% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nest2species}
\alias{nest2species}
\title{translates nest id in species abbreviation}
\usage{
nest2species(nest)
}
\arguments{
\item{nests}{nest ID in the format SXYY (S = one letter for the species, X = GPS ID, YY = running nest number)}
}
\value{
species abbreviation (REPH, RNPH or PESA)
}
\description{
this function extracts the species from the nest ID (R = REPH, N = RNPH. P = PESA)
}
\examples{
nest2species( nest = c('R101', 'P201', 'X102', 'N101')  )
}
