% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{nest_attendance_gps_tag}
\alias{nest_attendance_gps_tag}
\title{nest attendance based on gps tags data}
\usage{
nest_attendance_gps_tag(last_datetime = Sys.time() - 3600 * 24 * 5,
  buffer = 2)
}
\arguments{
\item{last_datetime}{last datetime. default to Sys.time() - 3600*6}

\item{buffer}{meters. passed to st_buffer(dist = )}

\item{project}{when TRUE, the default, data.tableTransform is called}
}
\description{
nest attendance based on gps tags data
}
