% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\name{NESTS2EGGS_CHICKS}
\alias{NESTS2EGGS_CHICKS}
\title{NESTS2EGGS_CHICKS
populate  table}
\usage{
NESTS2EGGS_CHICKS(Species = "REPH", table = "EGGS_CHICKS",
  db = "FIELD_REPHatBARROW")
}
\description{
NESTS2EGGS_CHICKS
populate  table
}
\examples{
NESTS2EGGS_CHICKS()
NESTS2EGGS_CHICKS(c('PESA', 'RNPH'), table = 'EGGS_CHICKS_field')
}
