% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_percent_light.R
\name{one_percent_light}
\alias{one_percent_light}
\title{Calculate the 1% Light Level}
\usage{
one_percent_light(depths, light)
}
\arguments{
\item{depths}{Vector of profile depths}

\item{light}{Vector of light measurements}
}
\description{
Calculate the 1% Light Level
}
