% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/success.R
\name{overall_success}
\alias{overall_success}
\title{Generate summaries of nest success from data}
\usage{
overall_success(
  path = get_default_data_path(),
  minyear = 1986,
  maxyear = as.integer(format(Sys.Date(), "\%Y")),
  download_if_missing = TRUE
)
}
\arguments{
\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}
}
\value{
a data.frame
}
\description{
Create a table of overall nest success by species and year
}
