% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_light.R
\name{per_light}
\alias{per_light}
\title{One Percent Light Level}
\usage{
per_light(depths, light)
}
\arguments{
\item{depths}{Vector of profile depths}

\item{do}{Vector of light measurements}
}
\description{
Calculate the one percent light level for a water profile
}
\examples{

per_light(depths, light)
}
\keyword{light}
\keyword{profile}
