% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_oxic.R
\name{percent_oxic}
\alias{percent_oxic}
\title{Calculate portion of water column that is oxic}
\usage{
percent_oxic(depths, do, digits)
}
\arguments{
\item{depths}{Vector of profile depths}

\item{do}{Vector of dissolved oxygen measurements}
}
\description{
Calculate portion of water column that is oxic
}
