% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_work.R
\name{plot_foraging}
\alias{plot_foraging}
\title{Plot tactile/visual foraging indicator data}
\usage{
plot_foraging(
  path = get_default_data_path(),
  minyear = 1986,
  maxyear = as.integer(format(Sys.Date(), "\%Y")),
  window = 3,
  download_if_missing = TRUE
)
}
\arguments{
\item{path}{either the file path that contains the Wading Bird folder or
"repo", which then pulls data from the EvergladesWadingBird GitHub repository}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{window}{number of years over which to create a rolling average}

\item{download_if_missing}{if the specified file path doesn't have the
EvergladesWadingBird folder, then download it}
}
\value{
a data.frame
}
\description{
Create a table of rolling averages for the proportion of tactile/visual foragers,
by year, and plot with thresholds
}
