% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_analyzer.R
\name{profile_analyzer}
\alias{profile_analyzer}
\title{Profile Analyzer}
\usage{
profile_analyzer(parameter)
}
\arguments{
\item{parameter}{Numeric vector}
}
\description{
Calculates summary statistics for a water profile parameter
}
\keyword{profile}
