% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Plot a depth profile for a given parameter}
\usage{
profile_plot(data, x, y, ...)
}
\arguments{
\item{data}{Data frame}

\item{x}{Variable to plot on x-axis}

\item{y}{Variable to plot on y-axis}

\item{...}{Optional ggplot2 arguments}
}
\value{
ggplot
}
\description{
Plot a depth profile for a given parameter
}
