% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{read_waypoints}
\alias{read_waypoints}
\title{read_waypoints}
\usage{
read_waypoints(x)
}
\description{
read_waypoints
}
\examples{
read_waypoints(x ='/media/mihai/GARMIN/Garmin/GPX/Waypoints_18-APR-17.gpx')
}
