% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{removeDuplicates
removeDuplicates (latest version on sdb)}
\usage{
removeDuplicates(table, key = "pk")
}
\description{
removeDuplicates
removeDuplicates (latest version on sdb)
}
\examples{
removeDuplicates('NESTS')
}
