% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.R
\name{reporter_NESTS}
\alias{reporter_NESTS}
\title{pdf reports}
\usage{
reporter_NESTS(Species)
}
\arguments{
\item{Species}{species}
}
\value{
a path to the pdf file
}
\description{
reporter_TABLENAME exports the TABLENAME summary as pdf
}
\examples{
reporter_NESTS('PESA')
}
