% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\name{reprojectBarrow}
\alias{reprojectBarrow}
\title{reprojectBarrow}
\usage{
reprojectBarrow(x, LAT = "lat", LON = "lon")
}
\description{
reprojectBarrow
}
\examples{
require(wadeR)
data("Barrow")
x = reprojectBarrow(Barrow)
}
