% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eden_covariates.R
\name{reshape_star}
\alias{reshape_star}
\title{Reshapes star object to dataframe}
\usage{
reshape_star(data, variable = "depth", year, boundaries)
}
\arguments{
\item{data}{data to reshape}

\item{variable}{value to reshape on}

\item{year}{data year}

\item{boundaries}{boundaries object}
}
\value{
data.frame depth data including columns for region, date, depth type value
}
\description{
Reshapes star object to dataframe
}
