% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\name{RESIGHTINGS_BY_ID}
\alias{RESIGHTINGS_BY_ID}
\title{RESIGHTINGS_BY_ID}
\usage{
RESIGHTINGS_BY_ID(UL, UR, LL, LR)
}
\description{
RESIGHTINGS_BY_ID
}
\examples{

RESIGHTINGS_BY_ID('M', 'G', 'Y,Y', 'Y')

}
