% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_work.R
\name{supercolony_indicator}
\alias{supercolony_indicator}
\title{Generate summaries of ibis supercolony indicator data}
\usage{
supercolony_indicator(
  path = get_default_data_path(),
  minyear = 1931,
  maxyear = 2021,
  window = 3,
  download_if_missing = TRUE
)
}
\arguments{
\item{path}{either the file path that contains the Wading Bird folder or
"repo", which then pulls data from the EvergladesWadingBird GitHub repository}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{window}{number of years over which to create a rolling average}

\item{download_if_missing}{if the specified file path doesn't have the
EvergladesWadingBird folder, then download it}
}
\value{
a data.frame
}
\description{
Create a table of rolling averages for the interval between ibis supercolony
events, by year
}
