% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermo_finder.R
\name{thermo_finder}
\alias{thermo_finder}
\title{Thermocline finder}
\usage{
thermo_finder(depths, temps, Smin = 0.1, seasonal = TRUE,
  index = FALSE, mixed_cutoff = 1)
}
\arguments{
\item{depths}{Vector of depths corresponding to the temps values}

\item{temps}{Vector of water temperature in degrees Celsius}

\item{Smin}{Optional parameter defining minimum density gradient for thermocline; defaults to 0.1}

\item{seasonal}{Logical value indicating if seasonal thermocline should be returned.
The seasonal thermocline is defined as the deepest density gradient found in the profile.}

\item{index}{Boolean value indicating if index of thermocline depth should be returned}

\item{mixed_cutoff}{Cutoff point where the thermoclien is not calculated; defaults to 1 degree C}
}
\description{
Determines the depth of the thermocline in a water column from a temperature profile
This function was borrowed heavily from the rLakeAnalyzer package on CRAN.
}
\examples{

depths <- c(0, 1, 2, 3, 4, 5, 6, 7, 8)
temps <- c(25, 25, 25, 20, 15, 12, 10, 8, 8, 8)

thermo_finder(temps, depths)
}
\keyword{thermocline}
