% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsi_calculator_chla.R
\name{tsi_calculator_chla}
\alias{tsi_calculator_chla}
\title{Trophic State Index - Chlorophyll-a}
\usage{
tsi_calculator_chla(chla_mgM3)
}
\arguments{
\item{chla_mgM3}{Numeric value of chlorophyll-a concentration in either ug/L or mg/m3}
}
\description{
Calculates the trophic state index value for a lake from a chloropyll-a concentration
}
\examples{

tsi_calculator_chla(0.10)
}
\keyword{state}
\keyword{trophic}
