% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsi_calculator_secchi.R
\name{tsi_calculator_secchi}
\alias{tsi_calculator_secchi}
\title{Trophic State Index - Secchi depth}
\usage{
tsi_calculator_secchi(secchi, units = "m")
}
\arguments{
\item{secchi}{Numeric value for a Secchi depth measurement}

\item{units}{Character string that accepts either 'm' or 'ft'; defaults to 'm' -- meters}
}
\description{
Calculates the trophic state index value for a lake from a Secchi depth measurement.
}
\examples{

tsi_calculator_secchi(1.5, units = "m")
}
\keyword{state}
\keyword{trophic}
