% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsi_calculator_tp.R
\name{tsi_calculator_tp}
\alias{tsi_calculator_tp}
\title{Trophic State Index - Total Phosphorus}
\usage{
tsi_calculator_tp(tp_epi, units = "mgL")
}
\arguments{
\item{tp_epi}{Numeric value of epilimnetic total phosphorus concentration}

\item{units}{Character string that accepts either ugL or mgL; defaults to mgL}
}
\description{
Calculates the trophic state index value for a lake from a epilimnetic total phosphorus concentration
}
\examples{

tsi_calculator_tp(0.01, units "mgL")
}
\keyword{state}
\keyword{trophic}
