% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsi_chla.R
\name{tsi_chla}
\alias{tsi_chla}
\title{Calculate the Carlson Trophic State Index - Chlorophyll-a}
\usage{
tsi_chla(chla, digits = 3)
}
\arguments{
\item{chla}{Chlorophyll-a concentration in ug/L}

\item{digits}{Rounding digits; defaults to 3}
}
\description{
Calculate the Carlson Trophic State Index - Chlorophyll-a
}
