% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsi_sd.R
\name{tsi_sd}
\alias{tsi_sd}
\title{Calculate the Carlson Trophic State Index - Secchi}
\usage{
tsi_sd(secchi, digits = 1)
}
\arguments{
\item{secchi}{Secchi depth in meters}

\item{digits}{Rounding digits; Defaults to 1}
}
\description{
Calculate the Carlson Trophic State Index - Secchi
}
