% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsi_tp.R
\name{tsi_tp}
\alias{tsi_tp}
\title{Calculate the Carlson Trophic State Index - Total Phosphorus}
\usage{
tsi_tp(tp, digits = 3)
}
\arguments{
\item{tp}{Epilimnetic total phosphorus concentration in ug/L}

\item{digits}{Rounding digits; Defaults to 3}
}
\description{
Calculate the Carlson Trophic State Index - Total Phosphorus
}
