% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{upload_gps}
\alias{upload_gps}
\title{upload_gps}
\usage{
upload_gps(p, gps_id)
}
\description{
upload_gps
}
\examples{
p = list.files('/media/mihai/GARMIN/Garmin', pattern = '.gpx$', full.names = TRUE, recursive = TRUE)
upload_gps(p)
}
