% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wader-package.R
\docType{package}
\name{wader}
\alias{wader}
\title{Creates summaries of the EvergladesWadingBird data}
\description{
This package is designed to be an interface to the Wading Bird data,
which resides online at \url{https://github.com/weecology/EvergladesWadingBird}.
It contains a set of functions to download, clean, and summarize the data.
}
\keyword{package}
