% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_column_oxic.R
\name{water_column_oxic}
\alias{water_column_oxic}
\title{Water Column Oxic - Percent}
\usage{
water_column_oxic(depths, do)
}
\arguments{
\item{depths}{Vector of profile depths}

\item{do}{Vector of profile dissolved oxygen concentrations}
}
\description{
This function calculates the percentage of a water column that is oxic.
}
\examples{
depths <- c(0, 1, 2, 3, 4, 5, 6, 7, 8)
do <- c(10.0, 12.3, 12.8, 9.1, 8.2, 4.1, 2.3, 1.2)

water_column_oxic(depths, do)
}
\keyword{profile}
