% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{water_density}
\alias{water_density}
\title{Water density}
\usage{
water_density(temps, salinity = temps * 0)
}
\arguments{
\item{temps}{Vector of water temperature in degrees Celsius}

\item{salinity}{Vector of salinity in Practical Salinity Units.}
}
\description{
Calculates density of water from temperature and salinity.
This function was borrowed heavily from the rLakeAnalyzer package on CRAN.
}
\examples{

water_density(temps, salinity)
}
\keyword{density}
