% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{water_report}
\alias{water_report}
\title{Plot water gauge data}
\usage{
water_report(
  depths,
  minyear = as.integer(format(Sys.Date(), "\%Y")) - 3,
  maxyear = as.integer(format(Sys.Date(), "\%Y")),
  wca = "1",
  download_if_missing = TRUE
)
}
\arguments{
\item{depths}{Depth data .csv file}

\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{wca}{wca to plot}

\item{download_if_missing}{if the specified file path doesn't have the
EvergladesWadingBird folder, then download it}
}
\value{
a data.frame
}
\description{
Create a plot of water gauge data
}
