% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{Weather by day or month}
\usage{
weather(
  level = "daily",
  horizon = 365,
  temperature_limit = 10,
  path = get_default_data_path()
)
}
\arguments{
\item{level}{specify 'monthly', 'daily'}

\item{horizon}{Horizon (number of days) to use when calculating cumulative values
(eg warm weather precip)}

\item{temperature_limit}{Temperature limit (in C) to use when calculating cumulative values
(eg warm weather precip)}

\item{path}{either the file path that contains the Wading Bird folder or
"repo", which then pulls data from the EvergladesWadingBird GitHub repository}
}
\description{
Summarize hourly weather data to either daily or monthly
}
