% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather_report}
\alias{weather_report}
\title{Weather figures for reporting}
\usage{
weather_report(
  data = "temperature",
  plot_year = as.numeric(format(Sys.Date(), "\%Y")),
  path = get_default_data_path()
)
}
\arguments{
\item{data}{specify 'temperature', 'precipitation'}

\item{plot_year}{year to plot}
}
\description{
Plot monthly deviations from mean
}
