% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yy2dbnam}
\alias{yy2dbnam}
\title{make field database name from year}
\usage{
yy2dbnam(year = format(Sys.Date(), format = "\%Y"),
  db = getOption("wader.db"))
}
\description{
make field database name from year
}
