% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_nul.R
\name{zoo_nul}
\alias{zoo_nul}
\title{Calculate zooplankton concentration from tow-net sampled}
\usage{
zoo_nul(count, tow_length, sample_vol, new_dia = 12.5)
}
\arguments{
\item{count}{Number of individuals counted}

\item{tow_length}{Length of tow in meters}

\item{sample_vol}{Sample volume in mL}

\item{new_dia}{Tow net diameter; defaults to 12.5}
}
\value{
A double
}
\description{
Calculate zooplankton concentration from tow-net sampled
}
