% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{inspector.CAPTURES}
\alias{inspector.CAPTURES}
\title{inspector functions
inspector functions as defined in DataEntry.validation package}
\usage{
\method{inspector}{CAPTURES}(x, ...)
}
\description{
inspector functions
inspector functions as defined in DataEntry.validation package
}
\examples{
\dontrun{
x = idbq('select * from EGGS_CHICKS ')
class(x) = c(class(x), 'EGGS_CHICKS')
ii = inspector(x)
evalidators(ii)
}

}
