% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectors.R
\name{inspector}
\alias{inspector}
\title{data inspector}
\usage{
inspector(x)
}
\arguments{
\item{x}{numeric number}

\item{...}{other arguments}
}
\description{
Data police
}
\details{
inspectors are a collection of validators
}
\examples{
1 # TODO


}
