% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_ui}
\alias{install_ui}
\alias{reboot_webserver}
\title{Install UI-s on localhost}
\usage{
install_ui(pwd, install_package = TRUE, root = "/srv/shiny-server",
  IP = ip())

reboot_webserver(pwd)
}
\arguments{
\item{pwd}{root pwd}

\item{root}{default to "/srv/shiny-server"}
}
\description{
Install UI-s on localhost
}
\examples{
install_ui()

reboot_webserver()

}
