% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_boundaries}
\alias{load_boundaries}
\title{Load ecohydrological boundaries}
\usage{
load_boundaries(
  path = file.path(get_default_data_path(), "EvergladesWadingBird",
    "SiteandMethods/regions"),
  level = "subregions"
)
}
\arguments{
\item{path}{path to regions shapefile}

\item{level}{region level to load (all, wcas, or subregions)}
}
\description{
Load ecohydrological boundaries
}
