% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_work.R
\name{max_counts}
\alias{max_counts}
\title{Generate summaries of max count data for target species and plot}
\usage{
max_counts(
  path = get_default_data_path(),
  minyear = 1980,
  maxyear = 2021,
  window = 3,
  download_if_missing = TRUE
)
}
\arguments{
\item{minyear}{Earliest year to include}

\item{maxyear}{Most recent year to include}

\item{window}{number of years over which to create a rolling average}
}
\value{
a data.frame
}
\description{
Create a table of rolling averages for the max count indicator data,
by year and region
}
