% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_percent_light.R
\name{one_percent_light}
\alias{one_percent_light}
\title{One Percent Light Level}
\usage{
one_percent_light(depths, light)
}
\arguments{
\item{depths}{Vector of profile depths}

\item{do}{Vector of light measurements}
}
\description{
Calculate the one percent light level for a water profile
}
\examples{

one_percent_light(depths, light)
}
\keyword{light}
\keyword{profile}
