% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_nul.R
\name{zoo_nul}
\alias{zoo_nul}
\title{Zooplankton NU  per liter}
\usage{
zoo_nul(df, towLen, sampleVol, netDia = 12.5)
}
\arguments{
\item{df}{Dataframe containing genus and totalCount of each genus}

\item{towLen}{Length in meters of zooplankton net tow at time of sampling}

\item{sampleVol}{Volume in mL of sample at time of identification}

\item{netDia}{Diameter in cm of zooplankton net opening; defaults to 12.5}
}
\description{
Calculates the natural units per liter of zooplankton from a sample
}
\examples{

zoo_nul(planktonCounts, towlength, sampleVol, netDia)
}
\keyword{zooplankton}
